requirements.txt
robotframework
robotframework-browser
robotframework-pabot
robotframework-requests
robotframework-jsonlibrary
robotframework-faker
dotenv
boto3